requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
absl-py==1.0.0
    # via mediapipe
attrs==21.4.0
    # via mediapipe
cycler==0.11.0
    # via matplotlib
fonttools==4.33.3
    # via matplotlib
keras==2.8.0
    # via -r requirements.in
kiwisolver==1.4.2
    # via matplotlib
matplotlib==3.5.2
    # via mediapipe
mediapipe==0.8.10
    # via -r requirements.in
numpy==1.22.3
    # via
    #   -r requirements.in
    #   matplotlib
    #   mediapipe
    #   opencv-contrib-python
opencv-contrib-python==4.5.5.64
    # via
    #   -r requirements.in
    #   mediapipe
packaging==21.3
    # via matplotlib
pillow==9.1.1
    # via
    #   -r requirements.in
    #   matplotlib
protobuf==3.20.1
    # via mediapipe
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via matplotlib
six==1.16.0
    # via
    #   absl-py
    #   python-dateutil
